+++ packages/editor-core/src/transformers/bitbucket/util.ts
@@ -1,6 +1,33 @@
-import { Node, DOMParser } from '@atlaskit/editor-core';
-import schema from './schema';
-import arrayFrom from './util/array-from';
+/**
+ * A replacement for `Array.from` until it becomes widely implemented.
+ */
+function arrayFrom(obj: any): any[] {
+  return Array.prototype.slice.call(obj);
+}
+
+/**
+ * A replacement for `String.repeat` until it becomes widely available.
+ */
+export function stringRepeat(text: string, length: number): string {
+  let result = '';
+  for (let x = 0; x < length; x++) {
+    result += text;
+  }
+  return result;
+}
+
+/**
+ * This function escapes all plain-text sequences that might get converted into markdown
+ * formatting by Bitbucket server (via python-markdown).
+ * @see MarkdownSerializerState.esc()
+ */
+export function escapeMarkdown(str: string, startOfLine?: boolean): string {
+  str = str.replace(/[`*\\+_]/g, '\\$&');
+  if (startOfLine) {
+    str = str.replace(/^[#-*]/, '\\$&').replace(/^(\d+)\./, '$1\\.');
+  }
+  return str;
+}
 
 /**
  * This function gets markup rendered by Bitbucket server and transforms it into markup that
@@ -113,12 +140,4 @@ export function transformHtml(html: string): HTMLElement {
   });
 
   return el;
-}
-
-/**
- * This function gets html string and parses it into ProseMirror Node.
- * Note that all unsupported elements will be discarded after parsing.
- */
-export function parseHtml(html: string): Node {
-  return DOMParser.fromSchema(schema).parse(transformHtml(html));
 }