+++ packages/media-viewer/stories/viewer-image-story.tsx
@@ -4,12 +4,13 @@ import { createStorybookContext, wideImageFileId, largeImageFileId } from '@atla
 import {ImageViewer} from '../src/3.0/viewers/image';
 
 const context = createStorybookContext();
+
 const badFileId = {
   ...wideImageFileId,
   id: 'badId'
-}
+};
 
-storiesOf('Image Viewer', {})
+storiesOf('Viewers: Image', {})
  .add('Image View', () => (
     <ImageViewer
       context={context}